fix refactorings
